[Fix/#135] 찜하기 본인확인 추가

+++ src/main/java/com/codevumc/codev_backend/service/co_studyheart/CoStudyHeartServiceImpl.java
@@ -8,6 +8,7 @@ import com.codevumc.codev_backend.service.ResponseService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.Objects;
 import java.util.Optional;
 
 @AllArgsConstructor
@@ -23,8 +24,10 @@ public class CoStudyHeartServiceImpl extends ResponseService implements CoStudyH
                 this.coStudyMapper.insertCoHeartOfStudy(co_email,co_studyId);
                 return setResponse(200, "Complete", "찜등록이 완료되었습니다.");
             }else{
-                this.coStudyMapper.deleteCoHeartOfStudy(co_email,co_studyId);
-                return setResponse(200, "Complete", "찜등록이 취소되었습니다.");
+                if(Objects.equals(coStudyMapper.getCoHeartOfStudyEmail(co_studyId), co_email)){
+                    this.coStudyMapper.deleteCoHeartOfStudy(co_email,co_studyId);
+                    return setResponse(200, "Complete", "찜등록이 취소되었습니다.");
+                }
             }
         }catch(Exception e){
             e.printStackTrace();
